+++ src/views/skillDetailPanel.ts
@@ -9,7 +9,7 @@ import { Skill, buildGitHubUrl } from '../types';
 import { InstalledSkillsTreeDataProvider } from './installedProvider';
 
 export class SkillDetailPanel {
-    public static readonly viewType = 'agentSkills.skillDetail';
+    public static readonly viewType = 'agentOrganizer.skillDetail';
     
     private static panels: Map<string, SkillDetailPanel> = new Map();
     
@@ -46,10 +46,10 @@ export class SkillDetailPanel {
             async (message) => {
                 switch (message.command) {
                     case 'install':
-                        await vscode.commands.executeCommand('agentSkills.install', this._skill);
+                        await vscode.commands.executeCommand('agentOrganizer.install', this._skill);
                         break;
                     case 'uninstall':
-                        await vscode.commands.executeCommand('agentSkills.uninstall', this._skill);
+                        await vscode.commands.executeCommand('agentOrganizer.uninstall', this._skill);
                         break;
                     case 'openExternal':
                         if (message.url) {